Move getCurrentSection out of viewGlobals
