+++ modules/viewGlobals.js
@@ -1,21 +1,10 @@
 'use strict';
 const formHelpers = require('./forms');
 
-function getCurrentSection(sectionId, pageId) {
-    const isHomepage = sectionId === 'toplevel' && pageId === 'home';
-    if (isHomepage) {
-        return 'toplevel';
-    } else if (sectionId !== 'toplevel') {
-        return sectionId;
-    }
-}
-
 function init(app) {
     const setViewGlobal = (name, value) => {
         app.get('engineEnv').addGlobal(name, value);
     };
-
-    setViewGlobal('getCurrentSection', getCurrentSection);
 
     // a global function for finding errors from a form array
     setViewGlobal('getFormErrorForField', (errorList, fieldName) => {
@@ -28,6 +17,5 @@ function init(app) {
 }
 
 module.exports = {
-    init,
-    getCurrentSection
+    init
 };
